small fixes

+++ Utils/save_video.m
@@ -15,7 +15,7 @@ end
 f = f - min(f(f>-Inf));
 sz = size(f);
 open(v)
-writeVideo(v, uint8(reshape(f, sz(1),sz(2),1,sz(3)) / max(f(:)) * 2^8));
+writeVideo(v, uint8(reshape(double(f), sz(1),sz(2),1,sz(3)) / max(double(f(:))) * 2^8));
 close(v)
 [~,cmdout] = system(['ffmpeg -y -i ' filename '.avi -c:v libx264 -preset slow -crf 18 ' filename '.mp4']);
 delete([filename '.avi'])
